Move auto-detect to server side with redirection
