+++ app/routes.js
@@ -4,7 +4,6 @@ import 'ngRoute';
 import * as viewAnnouncement from '~/views/frames/announcement';
 import * as viewSchedule     from '~/views/frames/schedule';
 import * as viewEventInfo    from '~/views/help/event-information';
-import * as viewAutoDetect   from '~/views/help/auto-detect';
 import * as viewAuthorization from '~/views/authorization';
 import templateLoading  from '~/views/help/loading.html';
 import templateNotConf  from '~/views/help/not-configured.html';
@@ -16,7 +15,6 @@ app.config(['extendedRouteProvider', '$locationProvider', function($routeProvide
 
     $routeProvider.
         when('/',                       { template :  templateLoading } ).
-        when('/current',                { ...castView(viewAutoDetect) } ).
 
         when('/help/not-configured',    { template : templateNotConf  } ).
         when('/help/event-information', { ...castView(viewEventInfo), controllerAs: 'eventInformationCtrl'} ).
