+++ package.json
@@ -22,6 +22,7 @@
     "ejs": "3.0.1",
     "express": "4.17.1",
     "http-proxy": "^1.18.1",
+    "ky": "^1.7.2",
     "morgan": "1.9.1"
   },
   "devDependencies": {
@@ -34,8 +35,7 @@
     "@rollup/plugin-node-resolve": "^15.2.3",
     "rollup": "^4.3.0",
     "rollup-plugin-dynamic-import-variables": "^1.1.0",
-    "rollup-plugin-string": "^3.0.0",
-    "rollup-plugin-terser": "^7.0.2"
+    "rollup-plugin-string": "^3.0.0"
   },
   "license": "MIT",
   "private": true,
