+++ rollup.config.js
@@ -7,7 +7,6 @@ import json                     from '@rollup/plugin-json';
 import commonjs                 from '@rollup/plugin-commonjs';
 import dynamicImportVariables   from 'rollup-plugin-dynamic-import-variables';
 import { string }               from "rollup-plugin-string";
-import { terser }               from 'rollup-plugin-terser';
 import bootWebApp, { cdnUrl }   from './app/boot.js';
 
 const isWatchOn = process.argv.includes('--watch');
